move the lock function to resource_models to avoid cyclical dependency
