+++ rest-service/manager_rest/storage/models.py
@@ -66,4 +66,5 @@ from .resource_models import (
     ExecutionSchedule,
     BlueprintLabel,
     DeploymentGroupLabel,
+    deployments_lock
 )
