+++ rest-service/manager_rest/storage/resource_models.py
@@ -58,7 +58,6 @@ from .relationships import (
     one_to_many_relationship,
     many_to_many_relationship
 )
-from manager_rest.storage import storage_utils
 
 if typing.TYPE_CHECKING:
     from manager_rest.resource_manager import ResourceManager
@@ -1234,7 +1233,7 @@ class ExecutionGroup(CreatedAtMixin, SQLResourceBase):
             if exc.status == ExecutionState.PENDING
         ]
         with sm.transaction():
-            storage_utils.deployments_lock()
+            deployments_lock()
             for execution in executions[self.concurrency:]:
                 execution.status = ExecutionState.QUEUED
                 sm.update(execution, modified_attrs=('status', ))
@@ -1953,3 +1952,19 @@ class DeploymentLabelsDependencies(BaseDeploymentDependencies):
     _source_deployment = foreign_key(Deployment._storage_id)
     _target_deployment = foreign_key(Deployment._storage_id)
 # endregion
+
+
+def deployments_lock():
+    """Lock multi-deployments update.
+
+    This lock is to be used around operations that possibly access multiple
+    deployments, their dependencies, and their executions.
+
+    This is not strictly based on deployments themselves, but to be used
+    co-operatively by functions, to avoid deadlocks in case multiple locks
+    on the deployments table are going to be acquired.
+    """
+    # the number doesn't mean anything, just needs to be globally unique,
+    # ie. not used by any other kind of lock across all of Cloudify
+    lock = 5
+    db.session.execute('SELECT pg_advisory_xact_lock(:lock)', {'lock': lock})
